Add status bullet to locale picker
