+++ packages/core/admin/package.json
@@ -40,8 +40,8 @@
     "@fortawesome/react-fontawesome": "^0.1.14",
     "@strapi/babel-plugin-switch-ee-ce": "1.0.0",
     "@strapi/helper-plugin": "3.6.8",
-    "@strapi/icons": "0.0.1-alpha.33",
-    "@strapi/parts": "0.0.1-alpha.33",
+    "@strapi/icons": "0.0.1-alpha.34",
+    "@strapi/parts": "0.0.1-alpha.34",
     "@strapi/utils": "3.6.8",
     "axios": "^0.21.1",
     "babel-loader": "8.2.2",
