+++ packages/core/helper-plugin/package.json
@@ -57,8 +57,8 @@
     "@storybook/builder-webpack5": "^6.3.7",
     "@storybook/manager-webpack5": "^6.3.7",
     "@storybook/react": "^6.3.7",
-    "@strapi/icons": "0.0.1-alpha.33",
-    "@strapi/parts": "0.0.1-alpha.33",
+    "@strapi/icons": "0.0.1-alpha.34",
+    "@strapi/parts": "0.0.1-alpha.34",
     "babel-loader": "^8.2.2",
     "cross-env": "^7.0.3",
     "enzyme": "^3.8.0",
