--- packages/plugins/i18n/admin/src/components/CMEditViewInjectedComponents/CMEditViewLocalePicker/Wrapper.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  padding-bottom: ${({ paddingBottom }) => paddingBottom};
-  border-top: 1px solid rgba(14, 22, 34, 0.04);
-`;
-
-export default Wrapper;
