+++ packages/plugins/i18n/admin/src/components/CMEditViewInjectedComponents/CMEditViewLocalePicker/utils/index.js
@@ -1,2 +1,1 @@
-export { default as addStatusColorToLocale } from './addStatusColorToLocale';
 export { default as createLocalesOption } from './createLocalesOption';
